feat(arc): modularize rust template
