+++ misc/arc-templates/arc_rust/build.ts
@@ -5,6 +5,9 @@ gulp.task("build", async () => {
     const options = { debug: true, cargo: true };
     const data = await Service.compileFiles([
         project.getFile("src/lib.rs"),
+        project.getFile("src/arc_module.rs")
+        project.getFile("src/color.rs"),
+        project.getFile("src/utils.rs"),
         project.getFile("Cargo.toml")            
     ], "rust", "wasm", options);
     const outWasm = project.newFile("out/lib.wasm", "wasm", true);
